net/mlx5: fix offset size in conntrack flow action
[ upstream commit 04aa27164533702ca5b9b00bd78aff39c816e7af ]

In mlx5 PMD, struct mlx5_aso_ct_action represents a single ASO CT object
inside a ASO CT bulk, which is used to implement conntrack flow action.
offset field, which is 16 bits wide, in this struct represents
an offset of the given object inside a bulk.

When SW Steering flow engine is used, ASO CT bulks are allocated on
demand, each having 64 objects. In this case offset field can fit any
valid object offset.

When HW Steering flow engine is used, a single ASO CT bulk is allocated
during flow engine configuration. Since HW supports bulks bigger
than 2^16 this offset field is too small to fit all valid values.
This causes some ASO CT objects to be incorrectly reused on multiple
connections, when more than 2^16 conntrack flow actions are used.

This patch fixes this behavior. offset field size is increased
to 32 bits.

Fixes: 463170a7c934 ("net/mlx5: support connection tracking with HWS")

Acked-by: Ori Kam <orika@nvidia.com>

+++ drivers/net/mlx5/mlx5.h
@@ -1245,7 +1245,7 @@ struct mlx5_aso_ct_action {
 	/* General action object for reply dir. */
 	void *dr_action_rply;
 	uint32_t refcnt; /* Action used count in device flows. */
-	uint16_t offset; /* Offset of ASO CT in DevX objects bulk. */
+	uint32_t offset; /* Offset of ASO CT in DevX objects bulk. */
 	uint16_t peer; /* The only peer port index could also use this CT. */
 	enum mlx5_aso_ct_state state; /* ASO CT state. */
 	bool is_original; /* The direction of the DR action to be used. */
